refactor: rework the file to make the action exportable
