+++ .github/workflows/main.yml
@@ -10,16 +10,12 @@ jobs:
   update_clickup_status:
     runs-on: ubuntu-latest
 
-    strategy:
-      matrix:
-        node-version: [12.x]
-
     steps:
     - uses: actions/checkout@v2
-    - name: Use Node.js ${{ matrix.node-version }}
+    - name: Use Node.js 12.x
       uses: actions/setup-node@v1
       with:
-        node-version: ${{ matrix.node-version }}
+        node-version: 12.x
     - run: yarn install
     - run: yarn run updateClickupStatus
       env:
